backend/pyproject.toml
[project]
name = "ebobot"
version = "0.1.0"
description = "Robot Psychology Hotline — art installation backend"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.9",
    "websockets>=13.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-playwright>=0.6.2",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["../tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
